Began some refactoring to allow adding/removing edges and hindrances. Realized it isn't going to work in the long term.
I'll have to adjust the way things work so that a person is a base
character with a stack of modifiers. Adding a new modifier puts it on
top of the stack; removing a modifier requires popping the stack to get
to it, removing it, and then reapplying the modifiers that were above
it.